Feat: Modification to the Health class
Health now inherits from ImageView class.

Remove the function to get path from an image and update the other
functions according to the change, integrated to the Tools class.

Implemented two new functions in favor of which a new attribute
'coordenates' was added. A function to search for a specific live and
another one to resset the orientation of the life image. More get and
set.