Feat: Modifications in loaderWindow and implementation of new functions.
All the necessary libraries were imported.

Implemented constants.

Implemented new buttons and things for the user.

Organized the code fo the initialize function by other functions
dividing the work:
- LoadSettings: load the data for the controller on screen.
- registerShips: register the boat and their lives in a dictionary-like
  data type.
- laodBoats: loads the boats in the vbox

Implemented function to return to the previus window.
Implemented a function to select boats.

Function to set boats, rotate boats,set random boats and delete boats
were created.
Also a function was created for when the game is started which changes
the window and loads the data to another controller when pressing start
game.